+++ packages/database/secondaryDbSchema/foreupBookingSaleItem.ts
@@ -1,6 +1,6 @@
 import type { InferSelectModel } from "drizzle-orm";
 import { sql } from "drizzle-orm";
-import { datetime, decimal, double, tinyint, unique, varchar } from "drizzle-orm/mysql-core";
+import { datetime, decimal, double, smallint, tinyint, unique, varchar } from "drizzle-orm/mysql-core";
 import { mySqlTable } from "./_table";
 
 export const foreupBookingSaleItem = mySqlTable(
@@ -15,7 +15,7 @@ export const foreupBookingSaleItem = mySqlTable(
     attributes_total: double("attributes_total"),
     attributes_subTotal: double("attributes_subTotal"),
     attributes_tax: double("attributes_tax"),
-    attributes_quantity: tinyint("attributes_quantity"),
+    attributes_quantity: smallint("attributes_quantity"),
     attributes_unitPrice: double("attributes_unitPrice"),
     attributes_unitCost: double("attributes_unitCost"),
     attributes_discountPercent: double("attributes_discountPercent"),
@@ -25,7 +25,7 @@ export const foreupBookingSaleItem = mySqlTable(
     attributes_price_category: varchar("attributes_price_category", { length: 36 }),
     attributes_type: varchar("attributes_type", { length: 15 }),
     attributes_holes: tinyint("attributes_holes"),
-    attributes_player_count: tinyint("attributes_player_count"),
+    attributes_player_count: smallint("attributes_player_count"),
     attributes_teesheet_name: varchar("attributes_teesheet_name", { length: 255 }),
     attributes_teetime_id: varchar("attributes_teetime_id", { length: 36 }),
     attributes_department: varchar("attributes_department", { length: 36 }),
